Revert "add libbpf asset (#752)"
This reverts commit 4c9c150e4bff4edb854be6f9647c2b963f69b339.
