+++ .github/workflows/image-base.yml
@@ -21,37 +21,28 @@ jobs:
             registry: quay.io/sustainable_computing_io
             username: ${{ secrets.BOT_NAME }}
             password: ${{ secrets.BOT_TOKEN }}
-      - name: Build and push bcc-amd64
+      - name: Build and push amd64
         uses: docker/build-push-action@v3
         with:
           context: .
-          file: ./build/Dockerfile.bcc.base
+          file: ./build/Dockerfile.base
           platforms: linux/amd64
           push: true
-          tags: quay.io/sustainable_computing_io/kepler_base:latest-bcc-amd64
-      - name: Build and push bcc-s390x
+          tags: quay.io/sustainable_computing_io/kepler_base:latest-amd64
+      - name: Build and push s390x
         uses: docker/build-push-action@v3
         with:
           context: .
-          file: ./build/Dockerfile.bcc.base.s390x
+          file: ./build/Dockerfile.base.s390x
           platforms: linux/s390x
           push: true
-          tags: quay.io/sustainable_computing_io/kepler_base:latest-bcc-s390x
-      - name: Build and push bcc-arm64
+          tags: quay.io/sustainable_computing_io/kepler_base:latest-s390x
         uses: docker/build-push-action@v3
         with:
           context: .
-          file: ./build/Dockerfile.bcc.base.arm64
+          file: ./build/Dockerfile.base.arm64
           platforms: linux/arm64
           push: true
-          tags: quay.io/sustainable_computing_io/kepler_base:latest-bcc-arm64
-      - name: Build and push libbpf-amd64
-        uses: docker/build-push-action@v3
-        with:
-          context: .
-          file: ./build/Dockerfile.libbpf.base
-          platforms: linux/amd64
-          push: true
-          tags: quay.io/sustainable_computing_io/kepler_base:latest-libbpf-amd64
+          tags: quay.io/sustainable_computing_io/kepler_base:latest-arm64
       - name: Create and upload multi-arch image
         run: make multi-arch-image-base
