+++ Makefile
@@ -56,67 +56,62 @@ else
 endif
 
 GENERAL_TAGS := 'include_gcs include_oss containers_image_openpgp gssapi providerless netgo osusergo gpu '
+BCC_TAG := 
+LIBBPF_TAG := 
+
+ifneq ($(shell command -v ldconfig),)
+	ifneq ($(shell ldconfig -p|grep bcc),)
+		BCC_TAG = bcc
+  	endif
+endif
+
+ifneq ($(shell command -v dpkg),)
+	ifneq ($(shell dpkg -l|grep bcc),)
+		BCC_TAG = bcc
+	endif
+endif
+
+ifneq ($(shell command -v ldconfig),)
+	ifneq ($(shell ldconfig -p|grep libbpf),)
+     LIBBPF_TAG = libbpf
+	endif
+endif
+
+ifneq ($(shell command -v dpkg),)
+	ifneq ($(shell dpkg -l|grep libbpf-dev),)
+		LIBBPF_TAG = libbpf
+	endif
+endif
+
 GO_LD_FLAGS := $(GC_FLAGS) -ldflags "-X $(LD_FLAGS)" $(CFLAGS)
 
 # set GOENV
 GOOS := $(shell go env GOOS)
 GOARCH := $(shell go env GOARCH)
+LIBBPF_HEADERS := /usr/include/bpf
+KEPLER_OBJ_SRC := $(SRC_ROOT)/bpfassets/libbpf/bpf.o/$(GOARCH)_kepler.bpf.o
+LIBBPF_OBJ := /usr/lib/$(ARCH)-linux-gnu/libbpf.a
+
 GOENV := GOOS=$(GOOS) GOARCH=$(GOARCH)
 
-ifdef ATTACHER_TAG
-	ATTACHER_TAG := $(ATTACHER_TAG)
-else
-# auto determine
-	BCC_TAG := 
-	LIBBPF_TAG := 
-
-	ifneq ($(shell command -v ldconfig),)
-		ifneq ($(shell ldconfig -p|grep bcc),)
-			BCC_TAG = bcc
-		endif
-	endif
-
-	ifneq ($(shell command -v dpkg),)
-		ifneq ($(shell dpkg -l|grep bcc),)
-			BCC_TAG = bcc
-		endif
-	endif
-
-	ifneq ($(shell command -v ldconfig),)
-		ifneq ($(shell ldconfig -p|grep libbpf),)
-		LIBBPF_TAG = libbpf
-		endif
-	endif
-
-	ifneq ($(shell command -v dpkg),)
-		ifneq ($(shell dpkg -l|grep libbpf-dev),)
-			LIBBPF_TAG = libbpf
-		endif
-	endif
-
-	LIBBPF_HEADERS := /usr/include/bpf
-	KEPLER_OBJ_SRC := $(SRC_ROOT)/bpfassets/libbpf/bpf.o/$(GOARCH)_kepler.bpf.o
-	LIBBPF_OBJ := /usr/lib/$(ARCH)-linux-gnu/libbpf.a
-
 # for libbpf tag, if libbpf.a, kepler.bpf.o exist, clear bcc tag
-	ifneq ($(LIBBPF_TAG),)
-		ifneq ($(wildcard $(LIBBPF_OBJ)),)
-			ifneq ($(wildcard $(KEPLER_OBJ_SRC)),)
-				BCC_TAG = 
-			endif
+ifneq ($(LIBBPF_TAG),)
+	ifneq ($(wildcard $(LIBBPF_OBJ)),)
+		ifneq ($(wildcard $(KEPLER_OBJ_SRC)),)
+			GOENV = GO111MODULE="" GOOS=$(GOOS) GOARCH=$(GOARCH) CGO_ENABLED=1 CC=clang CGO_CFLAGS="-I $(LIBBPF_HEADERS)" CGO_LDFLAGS="$(LIBBPF_OBJ)"
+			BCC_TAG = 
 		endif
 	endif
-# if bcc tag is not clear, clear libbpf tag
-	ifneq ($(BCC_TAG),)
-		LIBBPF_TAG = 
-	endif
-	ATTACHER_TAG := $(BCC_TAG)$(LIBBPF_TAG)
 endif
 
-# if libbpf tag is not empty, update goenv
-ifeq ($(ATTACHER_TAG),libbpf)
-	GOENV = GO111MODULE="" GOOS=$(GOOS) GOARCH=$(GOARCH) CGO_ENABLED=1 CC=clang CGO_CFLAGS="-I $(LIBBPF_HEADERS)" CGO_LDFLAGS="$(LIBBPF_OBJ)"
+# if bcc tag is not clear, clear libbpf tag
+ifneq ($(BCC_TAG),)
+	LIBBPF_TAG = 
 endif
+
+ATTACHER_TAG := $(BCC_TAG)$(LIBBPF_TAG)
+
+
 
 ifneq ($(ATTACHER_TAG),)
 	DOCKERFILE := $(SRC_ROOT)/build/Dockerfile.$(ATTACHER_TAG).kepler
@@ -299,6 +294,12 @@ set_govulncheck:
 govulncheck: set_govulncheck tidy-vendor
 	@govulncheck -v ./... || true
 
+e2e-libbpf:
+	$(GOENV) go test ./e2e/... --tags libbpf -v --race --bench=. -cover --count=1 --vet=all
+
+e2e-bcc:
+	go test ./e2e/... --tags bcc -v --race --bench=. -cover --count=1 --vet=all
+
 format:
 	./automation/presubmit-tests/gofmt.sh
 
@@ -344,7 +345,7 @@ cluster-up:
 .PHONY: cluster-up
 
 e2e:
-	./hack/verify.sh test ${ATTACHER_TAG}
+	./hack/verify.sh test
 .PHONY: e2e
 
 check: tidy-vendor set_govulncheck govulncheck format golint test
